feat(net): it receives.

+++ include/types.h
@@ -19,7 +19,7 @@ typedef unsigned char uchar;
 typedef unsigned char uint8;
 typedef unsigned short uint16;
 typedef unsigned int uint32;
-typedef unsigned long uint64;
+typedef unsigned long long uint64;
 
 typedef uint pde_t;
 typedef uint gfp_t;
